+++ packages/vchart-types/types/event/bubble.d.ts
@@ -9,7 +9,10 @@ export declare class Bubble {
     constructor();
     addHandler(handler: EventHandler<EventParams>, level: EventBubbleLevel): this;
     removeHandler(handler: EventHandler<EventParams>): this;
+    preventHandler(handler: EventHandler<EventParams>): this;
+    allowHandler(handler: EventHandler<EventParams>): this;
     getHandlers(level: EventBubbleLevel): EventHandler<EventParams>[];
+    getAllHandlers(): EventHandler<EventParams>[];
     getCount(): number;
     release(): void;
 }