+++ packages/vchart/src/core/vchart.ts
@@ -114,6 +114,7 @@ import {
   registerElementSelect as registerSelectInteraction
 } from '../interaction';
 import type { IIndicator } from '../component/indicator';
+import type { IGeoCoordinate } from '../component/geo';
 
 export class VChart implements IVChart {
   readonly id = createID();
@@ -343,8 +344,15 @@ export class VChart implements IVChart {
     const { dom, renderCanvas, mode, stage, poptip, ...restOptions } = this._option;
     const isTrueBrowseEnv = isTrueBrowser(mode);
 
+    // 根据 mode 配置动态加载浏览器或 node 环境代码
+    if (isTrueBrowseEnv) {
+      registerBrowserEnv();
+    } else if (mode === 'node') {
+      registerNodeEnv();
+    }
+
     if (isTrueBrowseEnv && dom) {
-      this._container = isString(dom) ? document?.getElementById(dom) : dom;
+      this._container = isString(dom) ? vglobal.getElementById(dom) : dom;
     }
     if (renderCanvas) {
       this._canvas = renderCanvas;
@@ -356,12 +364,6 @@ export class VChart implements IVChart {
     if (mode !== 'node' && !this._container && !this._canvas && !this._stage) {
       this._option?.onError('please specify container or renderCanvas!');
       return;
-    }
-    // 根据 mode 配置动态加载浏览器或 node 环境代码
-    if (isTrueBrowseEnv) {
-      registerBrowserEnv();
-    } else if (mode === 'node') {
-      registerNodeEnv();
     }
 
     this._viewBox = this._option.viewBox;
@@ -631,6 +633,7 @@ export class VChart implements IVChart {
       // 卸载了chart之后再设置主题 避免多余的reInit
       if (updateResult.changeTheme) {
         this._setCurrentTheme();
+        this._setFontFamilyTheme(this._currentTheme?.fontFamily as string);
       }
       // 如果不需要动画，那么释放item，避免元素残留
       this._compiler?.releaseGrammar(this._option?.animation === false || this._spec?.animation === false);
@@ -646,6 +649,7 @@ export class VChart implements IVChart {
       // 不remake的情况下，可以在这里更新主题
       if (updateResult.changeTheme) {
         this._setCurrentTheme();
+        this._setFontFamilyTheme(this._currentTheme?.fontFamily as string);
       }
       if (updateResult.reCompile) {
         // recompile
@@ -1802,7 +1806,7 @@ export class VChart implements IVChart {
     // 用户传入 canvas
     let canvasNode: Maybe<HTMLCanvasElement>;
     if (isString(this._canvas)) {
-      canvasNode = document?.getElementById(this._canvas) as HTMLCanvasElement;
+      canvasNode = vglobal.getElementById(this._canvas) as HTMLCanvasElement;
     } else {
       canvasNode = this._canvas as HTMLCanvasElement;
     }
@@ -2042,6 +2046,42 @@ export class VChart implements IVChart {
     const indicators = this._chart?.getComponentsByType(ComponentTypeEnum.indicator) as unknown as IIndicator[];
     if (indicators && indicators[index]) {
       indicators[index].updateDatum(datum);
+    }
+  }
+
+  /**
+   * 地图缩放 API
+   * @param [regionIndex=0] 根据索引顺序指定某个 region 区域的地图坐标系进行缩放
+   * @param zoom 缩放比例
+   * @param center 缩放中心
+   * @since 1.11.10
+   */
+  geoZoomByIndex(regionIndex: number = 0, zoom: number, center?: { x: number; y: number }) {
+    const region = this._chart?.getRegionsInQuerier({ regionIndex })[0];
+    const geoCoordinates = this._chart?.getComponentsByType(
+      ComponentTypeEnum.geoCoordinate
+    ) as unknown as IGeoCoordinate[];
+    const coord = geoCoordinates?.find(coord => coord.getRegions()?.includes(region));
+    if (coord) {
+      coord.dispatchZoom(zoom, center);
+    }
+  }
+
+  /**
+   * 地图缩放 API
+   * @param [regionId=0] 根据 region id 指定某个 region 区域的地图坐标系进行缩放
+   * @param zoom 缩放比例
+   * @param center 缩放中心
+   * @since 1.11.10
+   */
+  geoZoomById(regionId: string | number, zoom: number, center?: { x: number; y: number }) {
+    const region = this._chart?.getRegionsInQuerier({ regionId })[0];
+    const geoCoordinates = this._chart?.getComponentsByType(
+      ComponentTypeEnum.geoCoordinate
+    ) as unknown as IGeoCoordinate[];
+    const coord = geoCoordinates?.find(coord => coord.getRegions()?.includes(region));
+    if (coord) {
+      coord.dispatchZoom(zoom, center);
     }
   }