+++ packages/vchart/src/interaction/zoom/zoomable.ts
@@ -23,22 +23,18 @@ export interface ITriggerOption {
   delayTime: number;
   realTime: boolean;
 }
-export interface IZoomEventOptions {
-  shouldZoom?: boolean;
-  zoomCallback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void;
-  shouldScroll?: boolean;
-  scrollCallback?: (params: { scrollX: number; scrollY: number }, e: BaseEventParams['event']) => void;
-}
+
+export type ZoomEventParams = { zoomDelta: number; zoomX: number; zoomY: number };
+
+export type ZoomCallback = (params: ZoomEventParams, e: BaseEventParams['event']) => Record<string, any> | void;
+
 export interface IZoomable {
   initZoomable: (evt: IEvent, mode?: RenderMode) => void;
-  initZoomEventOfSeries: (
-    s: ISeries,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void
-  ) => any;
+  initZoomEventOfSeries: (s: ISeries, callback?: ZoomCallback) => any;
   initZoomEventOfRegions: (
     regions: IRegion[],
     filter?: (s: ISeries) => boolean,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void,
+    callback?: ZoomCallback,
     option?: ITriggerOption
   ) => void;
 
@@ -104,11 +100,7 @@ export class Zoomable implements IZoomable {
     return getDefaultTriggerEventByMode(this._renderMode)[type];
   }
 
-  private _zoomEventDispatch(
-    params: BaseEventParams,
-    regionOrSeries: IRegion | ISeries,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void
-  ) {
+  private _zoomEventDispatch(params: BaseEventParams, regionOrSeries: IRegion | ISeries, callback?: ZoomCallback) {
     if (!this._isGestureListener && !params.event) {
       return;
     }
@@ -131,14 +123,14 @@ export class Zoomable implements IZoomable {
     ) {
       return;
     }
-
+    let extendParams: ReturnType<typeof callback> = {};
     if (callback) {
-      // zoomDelta, zoomX, zoomY can be changed in the callback
-      callback({ zoomDelta, zoomX, zoomY }, event);
+      extendParams = callback({ zoomDelta, zoomX, zoomY }, event);
     }
     this._eventObj.emit('zoom', {
       scale: event.zoomDelta,
       scaleCenter: { x: event.zoomX, y: event.zoomY },
+      ...extendParams,
       model: this
     } as unknown as ExtendEventParam);
   }
@@ -159,7 +151,7 @@ export class Zoomable implements IZoomable {
   private _bindZoomEventAsRegion(
     eventObj: IEvent,
     regionOrSeries: IRegion | ISeries,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void,
+    callback?: ZoomCallback,
     option?: ITriggerOption
   ) {
     const delayType = option?.delayType ?? 'throttle';
@@ -171,8 +163,7 @@ export class Zoomable implements IZoomable {
       : [this._getZoomTriggerEvent('zoom'), { level: Event_Bubble_Level.chart, consume: true }];
     const zoomEndParams: [string] | [string, EventQuery] = this._isGestureListener
       ? [this._getZoomTriggerEvent('zoomEnd')]
-      : [this._getZoomTriggerEvent('zoomEnd'), { level: Event_Bubble_Level.chart, consume: true }];
-
+      : [this._getZoomTriggerEvent('zoomEnd'), { level: Event_Bubble_Level.chart, consume: false }];
     // pc端没有scrollEnd事件，所以漫游模式下scroll仅支持realTime
     (event as any).on(
       ...zoomEndParams,
@@ -191,11 +182,7 @@ export class Zoomable implements IZoomable {
     );
   }
 
-  initZoomEventOfSeries(
-    s: ISeries,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void,
-    option?: ITriggerOption
-  ) {
+  initZoomEventOfSeries(s: ISeries, callback?: ZoomCallback, option?: ITriggerOption) {
     if (this._option.disableTriggerEvent) {
       return;
     }
@@ -207,7 +194,7 @@ export class Zoomable implements IZoomable {
   initZoomEventOfRegions(
     regions: IRegion[],
     filter?: (s: ISeries) => boolean,
-    callback?: (params: { zoomDelta: number; zoomX: number; zoomY: number }, e: BaseEventParams['event']) => void,
+    callback?: ZoomCallback,
     option?: ITriggerOption
   ) {
     if (this._option.disableTriggerEvent) {
@@ -481,12 +468,16 @@ export class Zoomable implements IZoomable {
       this._zoomableTrigger.pointerId = null;
       this._eventObj.off(move, { level: Event_Bubble_Level.chart, source: Event_Source_Type.chart }, mousemove as any);
       this._eventObj.off(end, { level: Event_Bubble_Level.chart, source: Event_Source_Type.window }, mouseup as any);
+      this._eventObj.allow(end);
     }, delayTime);
+
     const mousemove = delayMap[delayType]((params: BaseEventParams) => {
       if (!this._zoomableTrigger.parserDragEvent(params.event)) {
         return;
       }
       this._clickEnable = false;
+      this._eventObj.prevent(end, mouseup as any);
+
       const event = params.event;
       const dx = event.canvasX - moveX;
       const dy = event.canvasY - moveY;