+++ packages/vchart/src/mark/base/base-line.ts
@@ -1,9 +1,10 @@
 import { isValidScaleType } from '@visactor/vscale';
 import type { StateValueType } from '../../compile/mark';
 import type { ConvertToMarkStyleSpec, ILineLikeMarkSpec } from '../../typings/visual';
-import { isFunction, isNil } from '@visactor/vutils';
-import { BaseMark } from './base-mark';
-import type { IMarkStyle, StyleConvert } from '../interface';
+import { isFunction, isNil, isValid } from '@visactor/vutils';
+import { BaseMark, ExChannelCall } from './base-mark';
+import type { IMarkOption, IMarkStyle, StyleConvert } from '../interface';
+import { Datum } from '../../typings';
 
 export abstract class BaseLineMark<T extends ILineLikeMarkSpec = ILineLikeMarkSpec> extends BaseMark<T> {
   protected abstract _getIgnoreAttributes(): string[];
@@ -62,11 +63,7 @@ export abstract class BaseLineMark<T extends ILineLikeMarkSpec = ILineLikeMarkSp
         enableSegments = true;
       }
 
-      let styleConverter = this._styleConvert(attrStyle);
-
-      if (isUserLevel && attr === 'angle') {
-        styleConverter = this.convertAngleToRadian(styleConverter as StyleConvert<number>);
-      }
+      const styleConverter = this._filterAttribute(attr as any, attrStyle, state, level, isUserLevel, stateStyle);
 
       this.setAttribute(attr as any, styleConverter, state, level, stateStyle);
     });