+++ packages/vchart/src/series/interface/type.ts
@@ -95,6 +95,7 @@ export const enum SeriesMarkNameEnum {
   barBackground = 'barBackground',
   lineLabel = 'lineLabel',
   areaLabel = 'areaLabel',
+  liquidGroup = 'liquidGroup',
   liquid = 'liquid',
   liquidBackground = 'liquidBackground',
   liquidOutline = 'liquidOutline',